Create Action For 2 and 3
Miled during this project was to create Action: 
 for “2. Terminate a process”:
The program will ask (receive user input integer) PID of the process to terminate. If the PID is valid
(process with that PID was created and exists), then the program kills the process with that PID;
otherwise, the program prints “no such PID” and goes back to the interface again. Refer to “man 2
kill” for more details of how to terminate a process.
Action for “3. Print active processes”:
The program should print all PIDs that were created and are still existing in the system. For your
convenience, assume that at most 10 child processes will exist at the same time. 

when running this code I simulated a menu and that allows the user to do those following processes .